moved api error inside appropriate module
