+++ server/src/server_data/endpoints/company_endpoints.rs
@@ -1,7 +1,7 @@
 use actix_web::web;
 use actix_web::{get, HttpResponse, Result};
 
-use crate::server_data::api_error::handle_api_result;
+use crate::server_data::endpoints::handle_api_result;
 use crate::server_data::repos::Repo;
 use crate::server_data::repos::company_repo::CompanyRepo;
