+++ server/src/server_data/endpoints/stonker_endpoints.rs
@@ -1,6 +1,6 @@
 use crate::models::stonker::NewStonker;
 use crate::repos::stonker_repo::StonkerRepo;
-use crate::server_data::api_error::handle_api_result;
+use crate::server_data::endpoints::handle_api_result;
 use crate::server_data::repos::Repo;
 use actix_web::web;
 use actix_web::{get, post, HttpResponse, Result};
