+++ server/src/server_data/mod.rs
@@ -2,4 +2,3 @@ pub mod endpoints;
 pub mod models;
 pub mod repos;
 pub mod schema;
-pub mod api_error;
